requirements.txt
fastapi-middleware==0.2.1
fastapi==0.115.8
hashids==1.3.1
libsql-experimental
pydantic==2.10.6
python-dotenv==1.0.1